Have k8s master nodes ignore updated image IDs
OS images are regularly updated on JetStream2, resulting in changing
image IDs. Thus, when running `terraform apply` on an existing cluster
(for example, to up/downscale the cluster), terraform recieves a
different image ID than the one the cluster was first created with,
prompting terraform to destroy/re-creates VMs. This can cause loss of
data, particularly if the k8s master nodes are affected.

Add the
[lifecycle](https://www.terraform.io/language/meta-arguments/lifecycle)
block to ensure changes to image IDs are ignored when updating k8s
master nodes.